third_party/blink/web_tests/fast/events/clientXY-in-zoom-and-scroll-expected.txt

Base
PASS event.clientX is 100
PASS event.clientY is 100
PASS event.pageX is 100
PASS event.pageY is 100

Just zoomed
FAIL event.clientX should be 83. Was 83.33332824707031.
FAIL event.clientY should be 83. Was 83.33332824707031.
FAIL event.pageX should be 83. Was 83.33332824707031.
FAIL event.pageY should be 83. Was 83.33332824707031.

Just scrolled
PASS event.clientX is 100
PASS event.clientY is 100
PASS event.pageX is 150
PASS event.pageY is 150

Zoomed and scrolled
FAIL event.clientX should be 83. Was 83.33332824707031.
FAIL event.clientY should be 83. Was 83.33332824707031.
FAIL event.pageX should be 133. Was 133.33332443237305.
FAIL event.pageY should be 133. Was 133.33332443237305.

RTL and scrolled
PASS event.clientX is 100
PASS event.clientY is 100
PASS event.pageX is 50
PASS event.pageY is 150
PASS successfullyParsed is true

TEST COMPLETE
